pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentctl"
version = "0.1.0"
description = "CLI for managing autonomous AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "google-cloud-compute>=1.14",
    "google-cloud-secret-manager>=2.16",
    "google-cloud-storage>=2.10",
    "google-api-python-client>=2.100",
    "google-auth>=2.20",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
agentctl = "agentctl.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
